Implement Insertion sort and update harness.
Let the user specify the sort algorithm via -t option. This is mapped
to an enum sort_algo_t for the supported algorithms.

Wrap everything in a sort_data_t so that the harness doesn't have to
change with each algorithm. It just calls the sort() function which
calls the appropriate function for the specified algorithm.

Add all sort related code into sort.c since there is no need to expose
the functions that implement the various algorithms.