Sequence, interleave and bundle an array of enumerables
sequence yeilds each element in the first enumerable one by one before
moving to the next enumerable and yields each element in that before
moving to the third and so on.

interleave yields the first element of the first enumerable, then the
first element of the second and so on until the last enumerable and
then loops to the the second element of the first enumerable and
repeats.

bundle yields an array of the first elements of all the enumerables,
then the second of all and so on up to the end of the smallest
enumerable.

bundle2 is similar to bundle but keeps going till the end of all the
enumerables.

Also implement heap, insertion, qsort and other ruby array, hash
practice work.

(Implemented about a week ago; committing it now.)