Sort digraph vertices in order of dependencies (ie topological sort)
Edge A->B means A is a pre-req of B or alternately B depends on A
which therefore should be done before B.

This is essentially a reverse post-order dfs traversal. Since a
topological sort only exists for DAGs, it first checks for any cycles
and, if none are found, then uses DepthFirstOrder to do the sort.

    $ ruby topological_sort.rb tinyDAG.txt
    Topological Sort: [8,7,2,3,0,5,1,6,9,11,10,12,4]

    $ ruby topological_sort.rb tinyDG.txt
    Graph is not a DAG
    Cycle: [[3,2,3], [3,5,4,2,3], [2,0,5,4,2], [8,6,8], [12,9,10,12]]