Use separate config for dev and production
These are stored in config/ directory and build on top of
webpack.common.config.js. Dev files are in dev/ directory which is
also where the dev bundle goes.

Passing the cli parameter env (one of dev or production) to yarn
builds the appropriate bundle. This defaults to dev.

Added webpack:prod to build the prod bundle. The existing webpack
script builds the dev bundle.